TMS_diplom/ecommerce/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-braces==1.15.0
django-jazzmin==2.5.0
django-oauth-toolkit==2.1.0
django-redis==5.2.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.1.0
idna==3.3
iniconfig==1.1.1
jwcrypto==1.3.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pip==22.2.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==63.1.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.12
wheel==0.37.1
wrapt==1.14.1
